Fixed recipe so it writes out each frame as a basic reduced image, rather than combining.
This required overwriting the output_file_name property in order to save each frame with a unique file name; for the time being, I use the original input frame name as part of the output file name. We should figure out what naming system we want to use for intermediate/final products, as this will be needed for many recipes.